feat(timeframes): implement POTTimeframe for time windows in `POTDetecto`
`POTTimeframe` has t0, t1, and t2 attributes which represent 3 time windows that are needed for the POT method. T0 is used to set the minimum period when calculating the POT Threshold via quantile, t1 is used to calculate the anomaly threshold via the quantile, and t2, which could be n% of dataset or 1 if in production mode, will be the "Time of Interest" for the user's usecase.