feat(detecto-model): implement fit with pot method (#7)
* fix(detectors): change predict method name into detect

* feat(models): add timeframes directory for different time windows classes

* feat(timeframes): create Timeframe abstract class as interface

Timeframe is an abstract class for other methods to divide time windows for statistic calculation that will be used to compare anomaly scores from the different time interval.

* feat(timeframes): implement POTTimeframe for time windows in `POTDetecto`

`POTTimeframe` has t0, t1, and t2 attributes which represent 3 time windows that are needed for the POT method. T0 is used to set the minimum period when calculating the POT Threshold via quantile, t1 is used to calculate the anomaly threshold via the quantile, and t2, which could be n% of dataset or 1 if in production mode, will be the "Time of Interest" for the user's usecase.

* feat(tests): create AbstractTimeframeTestModel for testing the Timeframe abstract class

* feat(tests): add timeframes directory for unit testing Timeframe classes

* fix(timeframes): rewrite **params into **kwargs as parameter in set_interval method

* fix(timeframes): rewrite **params into **kwargs in POTTimeframe set_interval method

* fix(tests): rewrite **params into **kwargs in AbstractTimeframeTestModel set_interval method

* test(timeframes): ensure correct implementation of `Timeframe` abstract class

* test(timeframes): ensure correct implementation of `POTTimeframe`

`POTTimeframe` is crucial for POT method in `POTDetecto`. Hence we need to ensure that the `set_interval()` method is correctly implemented.

* fix(tests): rewrite Detecto abstract method and its parameter

* feat(detectors): implement compute_pot_threshold and extract_exceedances

* test(detectors): implement compute_pot_threshold and extract_exceedances

Ensuring the correct implementation of calculation for POT Threshold and the extraction of its exceedances.

* docs(models): add Docstring to POTTimeframe

* fix(models): remove score abstract method from `Detecto` abstract class

* fix(models): add list to kwargs static type

* fix(tests): adjust methods for `AbstractDetectoTestModel`

* fix(models): add str to kwargs set_params method in `Detecto`

* fix(tests): add str to kwargs in set_params method

* fix(tests): adjust unit test for Detecto abstract class

Aligning the parameters in every method.

* fix(models): Adjust all params in all methods in `POTDetecto`

* fix(tests): rewrite unit tests for `pot` into `exceedance`

Adjusting the naming convention of the keyword `pot` into `exceedance`

* fix(static-type): rewrite the arrangement of static-type to str | int | float | None

* fix(tests): rewrite the test for testing the abstract `Detecto` class

* fix(static-type): change float into int type for t0, t1, t2

* feat(models): implement fit() method for POTDetecto with genpareto.fit()

This fitting method serves for extracting the c, scale, and loc parameters from genpareto fitting to calculate the p-value and the inverted p-value. The intverted p-value will be accumulated and form the "anomaly score" per feature and row+n.

* test(unit): ensure fit() method result in accumulated inverted p-value

The anomaly score is caclulated per row + n (continuous) and feature.

---------

Co-authored-by: N. L <nino@pleno.earth>