fix(detecto): handle inf in fit pot detecto (#9)
* fix(detector): hanlde 0 p-value as infinity

* test(unit): ensure correct infinity and None usage in fit method

If the list of fitting values is empty, than store None as the anomaly score per cell. If the result of survival function is 0 (p-value), then highly likely that it has avery small p-value that shows an extreme anomaly, thus store infinity.

* fix(detector): create more precise variable name for anomaly scores

Change `anomaly_score` to `total_anomaly_score_per_row` for the total accumulated inverted p-values.

* fix(detector): add total anomaly score per row as its own column in `fit()`

* test(unit): ensure `total_anomaly_score` column added to fit() result

* fix(detecto): rewrite set params method and the params structure

self.__params is a dictionary that contains the row indices for the first level of the keys and column names (features) in the second nested keys. THe actual items are: c, scale, loc, anomaly score per cell, total anomaly score per row => For each row and feature.

---------

Co-authored-by: N. L <nino@pleno.earth>