feat(detecto): implement anomaly detection method pot detecto (#16)
* feat(models): add attribute `anomaly_threshold` to `POTDetecto`

The attribute is to store the quantile of all anomaly scores above zero and not float("inf") from the `t1` timeframe.

* docs(models): add description for `anomaly_threshold` attribute in `POTDetecto`

* feat(models): create anomaly threshold calculation with quantile in T1 timeframe

* feat(models): add if statement for raising ValueError if no total anomaly scores > 0

* test(unit): ensure the computation of anomaly threshold is correct

The anomaly threshold is a single floating point number. This is calculated with quantile method where the values are the total anomaly scores from the T1 timeframe.

* feat(models): implement detect method

The detection method is actually a very simple way where we take all values from `T2` timeframe and compare them to the anomaly threshold. Whatever goes above the threshold is the anomaly.

* test(unit): ensure the correctness of the detection method

I sliped in a val * 100000 for the last element to see if it is detected as anomaly. And as expected, other high values within the `t2` timeframe are not seen as anomaly, only the last one as the full flow tries to get the extremely large values!

---------

Co-authored-by: N. L <nino@pleno.earth>